q-dev: device protocol
