+++ qubesadmin/devices.py
@@ -32,861 +32,12 @@ Devices are identified by pair of (backend domain, `ident`), where `ident` is
 :py:class:`str`.
 """
 import itertools
-import string
-import sys
+from typing import Optional, Iterable
+
 import qubesadmin
-from enum import Enum
-from typing import Optional, Dict, Any, List, Type, Iterable
-
-
-class UnexpectedDeviceProperty(qubesadmin.exc.QubesException, ValueError):
-    """
-    Device has unexpected property such as backend_domain, devclass etc.
-    """
-
-
-class ProtocolError(AssertionError):
-    """
-    Raised when something is wrong with data received.
-    """
-
-
-def qbool(value):
-    """
-    Property setter for boolean properties.
-
-    It accepts (case-insensitive) ``'0'``, ``'no'`` and ``false`` as
-    :py:obj:`False` and ``'1'``, ``'yes'`` and ``'true'`` as
-    :py:obj:`True`.
-    """
-
-    if isinstance(value, str):
-        lcvalue = value.lower()
-        if lcvalue in ('0', 'no', 'false', 'off'):
-            return False
-        if lcvalue in ('1', 'yes', 'true', 'on'):
-            return True
-        raise qubesadmin.exc.QubesValueError(
-            'Invalid literal for boolean property: {!r}'.format(value))
-
-    return bool(value)
-
-
-class Device:
-    def __init__(self, backend_domain, ident, devclass=None):
-        self.__backend_domain = backend_domain
-        self.__ident = ident
-        self.__bus = devclass
-
-    def __hash__(self):
-        return hash((str(self.backend_domain), self.ident))
-
-    def __eq__(self, other):
-        return (
-            self.backend_domain == other.backend_domain and
-            self.ident == other.ident
-        )
-
-    def __lt__(self, other):
-        if isinstance(other, Device):
-            return (self.backend_domain, self.ident) < \
-                   (other.backend_domain, other.ident)
-        return NotImplemented()
-
-    def __repr__(self):
-        return "[%s]:%s" % (self.backend_domain, self.ident)
-
-    def __str__(self):
-        return '{!s}:{!s}'.format(self.backend_domain, self.ident)
-
-    @property
-    def ident(self) -> str:
-        """
-        Immutable device identifier.
-
-        Unique for given domain and device type.
-        """
-        return self.__ident
-
-    @property
-    def backend_domain(self) -> 'qubesadmin.vm.QubesVM':
-        """ Which domain provides this device. (immutable)"""
-        return self.__backend_domain
-
-    @property
-    def devclass(self) -> str:
-        """ Immutable* Device class such like: 'usb', 'pci' etc.
-
-        For unknown devices "peripheral" is returned.
-
-        *see `@devclass.setter`
-        """
-        if self.__bus:
-            return self.__bus
-        else:
-            return "peripheral"
-
-    @property
-    def devclass_is_set(self) -> bool:
-        """
-        Returns true if devclass is already initialised.
-        """
-        return bool(self.__bus)
-
-    @devclass.setter
-    def devclass(self, devclass: str):
-        """ Once a value is set, it should not be overridden.
-
-        However, if it has not been set, i.e., the value is `None`,
-        we can override it."""
-        if self.__bus != None:
-            raise TypeError("Attribute devclass is immutable")
-        self.__bus = devclass
-
-
-class DeviceCategory(Enum):
-    """
-    Category of peripheral device.
-
-    Arbitrarily selected interfaces that are important to users,
-    thus deserving special recognition such as a custom icon, etc.
-    """
-    Other = "*******"
-
-    Communication = ("u02****", "p07****")  # eg. modems
-    Input = ("u03****", "p09****")  # HID etc.
-    Keyboard = ("u03**01", "p0900**")
-    Mouse = ("u03**02", "p0902**")
-    Printer = ("u07****",)
-    Scanner = ("p0903**",)
-    # Multimedia = Audio, Video, Displays etc.
-    Microphone = ("m******",)
-    Multimedia = ("u01****", "u0e****", "u06****", "u10****", "p03****",
-                  "p04****")
-    Wireless = ("ue0****", "p0d****")
-    Bluetooth = ("ue00101", "p0d11**")
-    Mass_Data = ("b******", "u08****", "p01****")
-    Network = ("p02****",)
-    Memory = ("p05****",)
-    PCI_Bridge = ("p06****",)
-    Docking_Station = ("p0a****",)
-    Processor = ("p0b****", "p40****")
-    PCI_Serial_Bus = ("p0c****",)
-    PCI_USB = ("p0c03**",)
-
-    @staticmethod
-    def from_str(interface_encoding: str) -> 'DeviceCategory':
-        result = DeviceCategory.Other
-        if len(interface_encoding) != len(DeviceCategory.Other.value):
-            return result
-        best_score = 0
-
-        for interface in DeviceCategory:
-            for pattern in interface.value:
-                score = 0
-                for t, p in zip(interface_encoding, pattern):
-                    if t == p:
-                        score += 1
-                    elif p != "*":
-                        score = -1  # inconsistent with pattern
-                        break
-
-                if score > best_score:
-                    best_score = score
-                    result = interface
-
-        return result
-
-
-class DeviceInterface:
-    """
-    Peripheral device interface wrapper.
-    """
-
-    def __init__(self, interface_encoding: str, devclass: Optional[str] = None):
-        ifc_padded = interface_encoding.ljust(6, '*')
-        if devclass:
-            if len(ifc_padded) > 6:
-                print(
-                    f"{interface_encoding=} is too long "
-                    f"(is {len(interface_encoding)}, expected max. 6) "
-                    f"for given {devclass=}",
-                    file=sys.stderr
-                )
-            ifc_full = devclass[0] + ifc_padded
-        else:
-            known_devclasses = {
-                'p': 'pci', 'u': 'usb', 'b': 'block', 'm': 'mic'}
-            devclass = known_devclasses.get(interface_encoding[0], None)
-            if len(ifc_padded) > 7:
-                print(
-                    f"{interface_encoding=} is too long "
-                    f"(is {len(interface_encoding)}, expected max. 7)",
-                    file=sys.stderr
-                )
-                ifc_full = ifc_padded
-            elif len(ifc_padded) == 6:
-                ifc_full = '?' + ifc_padded
-            else:
-                ifc_full = ifc_padded
-
-        self._devclass = devclass
-        self._interface_encoding = ifc_full
-        self._category = DeviceCategory.from_str(self._interface_encoding)
-
-    @property
-    def devclass(self) -> Optional[str]:
-        """ Immutable Device class such like: 'usb', 'pci' etc. """
-        return self._devclass
-
-    @property
-    def category(self) -> DeviceCategory:
-        """ Immutable Device category such like: 'Mouse', 'Mass_Data' etc. """
-        return self._category
-
-    @classmethod
-    def unknown(cls) -> 'DeviceInterface':
-        """ Value for unknown device interface. """
-        return cls("?******")
-
-    def __repr__(self):
-        return self._interface_encoding
-
-    def __str__(self):
-        if self.devclass == "block":
-            return "Block device"
-        if self.devclass in ("usb", "pci"):
-            result = self._load_classes(self.devclass).get(
-                self._interface_encoding[1:], None)
-            if result is None:
-                result = self._load_classes(self.devclass).get(
-                    self._interface_encoding[1:-2] + '**', None)
-            if result is None:
-                result = self._load_classes(self.devclass).get(
-                    self._interface_encoding[1:-4] + '****', None)
-            if result is None:
-                result = f"Unclassified {self.devclass} device"
-            return result
-        if self.devclass == 'mic':
-            return "Microphone"
-        return repr(self)
-
-    @staticmethod
-    def _load_classes(bus: str):
-        """
-        List of known device classes, subclasses and programming interfaces.
-        """
-        # Syntax:
-        # C class       class_name
-        #       subclass        subclass_name           <-- single tab
-        #               prog-if  prog-if_name   <-- two tabs
-        result = {}
-        with open(f'/usr/share/hwdata/{bus}.ids',
-                  encoding='utf-8', errors='ignore') as pciids:
-            class_id = None
-            subclass_id = None
-            for line in pciids.readlines():
-                line = line.rstrip()
-                if line.startswith('\t\t') and class_id and subclass_id:
-                    (progif_id, _, progif_name) = line[2:].split(' ', 2)
-                    result[class_id + subclass_id + progif_id] = \
-                        f"{class_name}: {subclass_name} ({progif_name})"
-                elif line.startswith('\t') and class_id:
-                    (subclass_id, _, subclass_name) = line[1:].split(' ', 2)
-                    # store both prog-if specific entry and generic one
-                    result[class_id + subclass_id + '**'] = \
-                        f"{class_name}: {subclass_name}"
-                elif line.startswith('C '):
-                    (_, class_id, _, class_name) = line.split(' ', 3)
-                    result[class_id + '****'] = class_name
-                    subclass_id = None
-
-        return result
-
-
-class DeviceInfo(Device):
-    """ Holds all information about a device """
-
-    def __init__(
-            self,
-            backend_domain: 'qubes.vm.BaseVM',
-            ident: str,
-            devclass: Optional[str] = None,
-            vendor: Optional[str] = None,
-            product: Optional[str] = None,
-            manufacturer: Optional[str] = None,
-            name: Optional[str] = None,
-            serial: Optional[str] = None,
-            interfaces: Optional[List[DeviceInterface]] = None,
-            parent: Optional[Device] = None,
-            attachment: Optional['qubes.vm.BaseVM'] = None,
-            self_identity: Optional[str] = None,
-            **kwargs
-    ):
-        super().__init__(backend_domain, ident, devclass)
-
-        self._vendor = vendor
-        self._product = product
-        self._manufacturer = manufacturer
-        self._name = name
-        self._serial = serial
-        self._interfaces = interfaces
-        self._parent = parent
-        self._attachment = attachment
-        self._self_identity = self_identity
-
-        self.data = kwargs
-
-    @property
-    def vendor(self) -> str:
-        """
-        Device vendor name from local database.
-
-        Could be empty string or "unknown".
-
-        Override this method to return proper name from `/usr/share/hwdata/*`.
-        """
-        if not self._vendor:
-            return "unknown"
-        return self._vendor
-
-    @property
-    def product(self) -> str:
-        """
-        Device name from local database.
-
-        Could be empty string or "unknown".
-
-        Override this method to return proper name from `/usr/share/hwdata/*`.
-        """
-        if not self._product:
-            return "unknown"
-        return self._product
-
-    @property
-    def manufacturer(self) -> str:
-        """
-        The name of the manufacturer of the device introduced by device itself.
-
-        Could be empty string or "unknown".
-
-        Override this method to return proper name directly from device itself.
-        """
-        if not self._manufacturer:
-            return "unknown"
-        return self._manufacturer
-
-    @property
-    def name(self) -> str:
-        """
-        The name of the device it introduced itself with.
-
-        Could be empty string or "unknown".
-
-        Override this method to return proper name directly from device itself.
-        """
-        if not self._name:
-            return "unknown"
-        return self._name
-
-    @property
-    def serial(self) -> str:
-        """
-        The serial number of the device it introduced itself with.
-
-        Could be empty string or "unknown".
-
-        Override this method to return proper name directly from device itself.
-        """
-        if not self._serial:
-            return "unknown"
-        return self._serial
-
-    @property
-    def description(self) -> str:
-        """
-        Short human-readable description.
-
-        For unknown device returns `unknown device (unknown vendor)`.
-        For unknown USB device returns `unknown usb device (unknown vendor)`.
-        For unknown USB device with known serial number returns
-            `<serial> (unknown vendor)`.
-        """
-        if self.product and self.product != "unknown":
-            prod = self.product
-        elif self.name and self.name != "unknown":
-            prod = self.name
-        elif self.serial and self.serial != "unknown":
-            prod = self.serial
-        elif self.parent_device is not None:
-            return f"sub-device of {self.parent_device}"
-        else:
-            prod = f"unknown {self.devclass if self.devclass else ''} device"
-
-        if self.vendor and self.vendor != "unknown":
-            vendor = self.vendor
-        elif self.manufacturer and self.manufacturer != "unknown":
-            vendor = self.manufacturer
-        else:
-            vendor = "unknown vendor"
-
-        return f"{prod} ({vendor})"
-
-    @property
-    def interfaces(self) -> List[DeviceInterface]:
-        """
-        Non-empty list of device interfaces.
-
-        Every device should have at least one interface.
-        """
-        if not self._interfaces:
-            return [DeviceInterface.unknown()]
-        return self._interfaces
-
-    @property
-    def parent_device(self) -> Optional[Device]:
-        """
-        The parent device if any.
-
-        If the device is part of another device (e.g. it's a single
-        partition of an usb stick), the parent device id should be here.
-        """
-        return self._parent
-
-    @property
-    def subdevices(self) -> List['DeviceInfo']:
-        """
-        The list of children devices if any.
-
-        If the device has subdevices (e.g. partitions of an usb stick),
-        the subdevices id should be here.
-        """
-        return [dev for dev in self.backend_domain.devices[self.devclass]
-                if dev.parent_device.ident == self.ident]
-
-    @property
-    def attachment(self) -> Optional['qubes.vm.BaseVM']:
-        """
-        VM to which device is attached (frontend domain).
-        """
-        return self._attachment
-
-    def serialize(self) -> bytes:
-        """
-        Serialize object to be transmitted via Qubes API.
-        """
-        # 'backend_domain', 'attachment', 'interfaces', 'data', 'parent_device'
-        # are not string, so they need special treatment
-        default_attrs = {
-            'ident', 'devclass', 'vendor', 'product', 'manufacturer', 'name',
-            'serial', 'self_identity'}
-        properties = b' '.join(
-            f'{prop}={serialize_str(value)}'.encode('ascii')
-            for prop, value in (
-                (key, getattr(self, key)) for key in default_attrs)
-        )
-
-        qname = serialize_str(self.backend_domain.name)
-        backend_prop = (b"backend_domain=" + qname.encode('ascii'))
-        properties += b' ' + backend_prop
-
-        if self.attachment:
-            qname = serialize_str(self.attachment.name)
-            attachment_prop = (b"attachment=" + qname.encode('ascii'))
-            properties += b' ' + attachment_prop
-
-        interfaces = serialize_str(
-            ''.join(repr(ifc) for ifc in self.interfaces))
-        interfaces_prop = (b'interfaces=' + interfaces.encode('ascii'))
-        properties += b' ' + interfaces_prop
-
-        if self.parent_device is not None:
-            ident = serialize_str(self.parent_device.ident)
-            ident_prop = (b'parent_ident=' + ident.encode('ascii'))
-            properties += b' ' + ident_prop
-            devclass = serialize_str(self.parent_device.devclass)
-            devclass_prop = (b'parent_devclass=' + devclass.encode('ascii'))
-            properties += b' ' + devclass_prop
-
-        data = b' '.join(
-            f'_{prop}={serialize_str(value)}'.encode('ascii')
-            for prop, value in ((key, self.data[key]) for key in self.data)
-        )
-        if data:
-            properties += b' ' + data
-
-        return properties
-
-    @classmethod
-    def deserialize(
-            cls,
-            serialization: bytes,
-            expected_backend_domain: 'qubes.vm.BaseVM',
-            expected_devclass: Optional[str] = None,
-    ) -> 'DeviceInfo':
-        try:
-            result = DeviceInfo._deserialize(
-                cls, serialization, expected_backend_domain, expected_devclass)
-        except Exception as exc:
-            print(exc, file=sys.stderr)
-            ident = serialization.split(b' ')[0].decode(
-                'ascii', errors='ignore')
-            result = UnknownDevice(
-                backend_domain=expected_backend_domain,
-                ident=ident,
-                devclass=expected_devclass,
-            )
-        return result
-
-    @staticmethod
-    def _deserialize(
-            cls: Type,
-            serialization: bytes,
-            expected_backend_domain: 'qubes.vm.BaseVM',
-            expected_devclass: Optional[str] = None,
-    ) -> 'DeviceInfo':
-        decoded = serialization.decode('ascii', errors='ignore')
-        ident, _, rest = decoded.partition(' ')
-        keys = []
-        values = []
-        key, _, rest = rest.partition("='")
-        keys.append(key)
-        while "='" in rest:
-            value_key, _, rest = rest.partition("='")
-            value, _, key = value_key.rpartition("' ")
-            values.append(deserialize_str(value))
-            keys.append(key)
-        value = rest[:-1]  # ending '
-        values.append(deserialize_str(value))
-
-        properties = dict()
-        for key, value in zip(keys, values):
-            if key.startswith("_"):
-                # it's handled in cls.__init__
-                properties[key[1:]] = value
-            else:
-                properties[key] = value
-
-        if properties['backend_domain'] != expected_backend_domain.name:
-            raise UnexpectedDeviceProperty(
-                f"Got device exposed by {properties['backend_domain']}"
-                f"when expected devices from {expected_backend_domain.name}.")
-        properties['backend_domain'] = expected_backend_domain
-
-        if 'attachment' not in properties or not properties['attachment']:
-            properties['attachment'] = None
-        else:
-            app = expected_backend_domain.app
-            properties['attachment'] = app.domains.get_blind(
-                properties['attachment'])
-
-        if expected_devclass and properties['devclass'] != expected_devclass:
-            raise UnexpectedDeviceProperty(
-                f"Got {properties['devclass']} device "
-                f"when expected {expected_devclass}.")
-
-        if properties["ident"] != ident:
-            raise UnexpectedDeviceProperty(
-                f"Got device with id: {properties['ident']} "
-                f"when expected id: {ident}.")
-
-        interfaces = properties['interfaces']
-        interfaces = [
-            DeviceInterface(interfaces[i:i + 7])
-            for i in range(0, len(interfaces), 7)]
-        properties['interfaces'] = interfaces
-
-        if 'parent_ident' in properties:
-            properties['parent'] = Device(
-                backend_domain=expected_backend_domain,
-                ident=properties['parent_ident'],
-                devclass=properties['parent_devclass'],
-            )
-            del properties['parent_ident']
-            del properties['parent_devclass']
-
-        return cls(**properties)
-
-    @property
-    def self_identity(self) -> str:
-        """
-        Get additional identification of device presented by device itself.
-
-        For pci/usb we expect:
-        <vendor_id>:<product_id>:<serial if any>:<interface1interface2...>
-        For block devices:
-        <parent_ident>:<interface number if any>
-
-        In addition to the description returns presented interfaces.
-        It is used to auto-attach usb devices, so an attacking device needs to
-        mimic not only a name, but also interfaces of trusted device (and have
-        to be plugged to the same port). For a common user it is all the data
-        she uses to recognize the device.
-        """
-        if not self._self_identity:
-            return "unknown:unknown:unknown:?******"
-        return self._self_identity
-
-
-def serialize_str(value: str):
-    return repr(str(value))
-
-
-def deserialize_str(value: str):
-    return value.replace("\\\'", "'")
-
-
-def sanitize_str(
-        untrusted_value: str,
-        allowed_chars: str,
-        replace_char: str = None,
-        error_message: str = ""
-) -> str:
-    """
-    Sanitize given untrusted string.
-
-    If `replace_char` is not None, ignore `error_message` and replace invalid
-    characters with the string.
-    """
-    if replace_char is None:
-        if any(x not in allowed_chars for x in untrusted_value):
-            raise ProtocolError(error_message)
-        return untrusted_value
-    result = ""
-    for char in untrusted_value:
-        if char in allowed_chars:
-            result += char
-        else:
-            result += replace_char
-    return result
-
-
-class UnknownDevice(DeviceInfo):
-    # pylint: disable=too-few-public-methods
-    """Unknown device - for example exposed by domain not running currently"""
-
-    def __init__(self, backend_domain, devclass, ident, **kwargs):
-        super().__init__(backend_domain, ident, devclass=devclass, **kwargs)
-
-
-class DeviceAssignment(Device):
-    """ Maps a device to a frontend_domain. """
-
-    def __init__(self, backend_domain, ident, options=None,
-                 frontend_domain=None, devclass=None,
-                 required=False, attach_automatically=False):
-        super().__init__(backend_domain, ident, devclass)
-        self.__options = options or {}
-        if required:
-            assert attach_automatically
-        self.__required = required
-        self.__attach_automatically = attach_automatically
-        self.__frontend_domain = frontend_domain
-
-    def clone(self, **kwargs):
-        """
-        Clone object and substitute attributes with explicitly given.
-        """
-        attr = {
-            "backend_domain": self.backend_domain,
-            "ident": self.ident,
-            "options": self.options,
-            "required": self.required,
-            "attach_automatically": self.attach_automatically,
-            "frontend_domain": self.frontend_domain,
-            "devclass": self.devclass,
-        }
-        attr.update(kwargs)
-        return self.__class__(**attr)
-
-    @classmethod
-    def from_device(cls, device: Device, **kwargs) -> 'DeviceAssignment':
-        """
-        Get assignment of the device.
-        """
-        return cls(
-            backend_domain=device.backend_domain,
-            ident=device.ident,
-            devclass=device.devclass,
-            **kwargs
-        )
-
-    @property
-    def device(self) -> DeviceInfo:
-        """Get DeviceInfo object corresponding to this DeviceAssignment"""
-        return self.backend_domain.devices[self.devclass][self.ident]
-
-    @property
-    def frontend_domain(self) -> Optional['qubesadmin.vm.QubesVM']:
-        """ Which domain the device is attached to. """
-        return self.__frontend_domain
-
-    @frontend_domain.setter
-    def frontend_domain(
-            self, frontend_domain: Optional['qubesadmin.vm.QubesVM']
-    ):
-        """ Which domain the device is attached to. """
-        self.__frontend_domain = frontend_domain
-
-    @property
-    def attached(self) -> bool:
-        """
-        Is the device attached to the fronted domain?
-
-        Returns False if device is attached to different domain
-        """
-        return self.device.attachment == self.frontend_domain
-
-    @property
-    def required(self) -> bool:
-        """
-        Is the presence of this device required for the domain to start? If yes,
-        it will be attached automatically.
-        """
-        return self.__required
-
-    @required.setter
-    def required(self, required: bool):
-        self.__required = required
-
-    @property
-    def attach_automatically(self) -> bool:
-        """
-        Should this device automatically connect to the frontend domain when
-        available and not connected to other qubes?
-        """
-        return self.__attach_automatically
-
-    @attach_automatically.setter
-    def attach_automatically(self, attach_automatically: bool):
-        self.__attach_automatically = attach_automatically
-
-    @property
-    def options(self) -> Dict[str, Any]:
-        """ Device options (same as in the legacy API). """
-        return self.__options
-
-    @options.setter
-    def options(self, options: Optional[Dict[str, Any]]):
-        """ Device options (same as in the legacy API). """
-        self.__options = options or {}
-
-    def serialize(self) -> bytes:
-        properties = b' '.join(
-            f'{prop}={serialize_str(value)}'.encode('ascii')
-            for prop, value in (
-                ('required', 'yes' if self.required else 'no'),
-                ('attach_automatically',
-                 'yes' if self.attach_automatically else 'no'),
-                ('ident', self.ident),
-                ('devclass', self.devclass)
-            )
-        )
-
-        back_name = serialize_str(self.backend_domain.name)
-        backend_domain_prop = (b"backend_domain=" + back_name.encode('ascii'))
-        properties += b' ' + backend_domain_prop
-
-        if self.frontend_domain is not None:
-            front_name = serialize_str(self.frontend_domain.name)
-            frontend_domain_prop = (
-                    b"frontend_domain=" + front_name.encode('ascii'))
-            properties += b' ' + frontend_domain_prop
-
-        if self.options:
-            properties += b' ' + b' '.join(
-                f'_{prop}={serialize_str(value)}'.encode('ascii')
-                for prop, value in self.options.items()
-            )
-
-        return properties
-
-    @classmethod
-    def deserialize(
-            cls,
-            serialization: bytes,
-            expected_backend_domain: 'qubes.vm.BaseVM',
-            expected_ident: str,
-            expected_devclass: Optional[str] = None,
-    ) -> 'DeviceAssignment':
-        try:
-            result = DeviceAssignment._deserialize(
-                cls, serialization,
-                expected_backend_domain, expected_ident, expected_devclass
-            )
-        except Exception as exc:
-            raise ProtocolError() from exc
-        return result
-
-    @staticmethod
-    def _deserialize(
-            cls: Type,
-            untrusted_serialization: bytes,
-            expected_backend_domain: 'qubes.vm.BaseVM',
-            expected_ident: str,
-            expected_devclass: Optional[str] = None,
-    ) -> 'DeviceAssignment':
-        options = {}
-        allowed_chars_key = string.digits + string.ascii_letters + '-_.'
-        allowed_chars_value = allowed_chars_key + ',+:'
-
-        untrusted_decoded = untrusted_serialization.decode(
-            'ascii', 'strict').strip()
-        keys = []
-        values = []
-        untrusted_key, _, untrusted_rest = untrusted_decoded.partition("='")
-
-        key = sanitize_str(
-            untrusted_key, allowed_chars_key,
-            error_message='Invalid chars in property name')
-        keys.append(key)
-        while "='" in untrusted_rest:
-            ut_value_key, _, untrusted_rest = untrusted_rest.partition("='")
-            untrusted_value, _, untrusted_key = ut_value_key.rpartition("' ")
-            value = sanitize_str(
-                deserialize_str(untrusted_value), allowed_chars_value,
-                error_message='Invalid chars in property value')
-            values.append(value)
-            key = sanitize_str(
-                untrusted_key, allowed_chars_key,
-                error_message='Invalid chars in property name')
-            keys.append(key)
-        untrusted_value = untrusted_rest[:-1]  # ending '
-        value = sanitize_str(
-            deserialize_str(untrusted_value), allowed_chars_value,
-            error_message='Invalid chars in property value')
-        values.append(value)
-
-        properties = dict()
-        for key, value in zip(keys, values):
-            if key.startswith("_"):
-                options[key[1:]] = value
-            else:
-                properties[key] = value
-
-        properties['options'] = options
-
-        if properties['backend_domain'] != expected_backend_domain.name:
-            raise UnexpectedDeviceProperty(
-                f"Got device exposed by {properties['backend_domain']} "
-                f"when expected devices from {expected_backend_domain.name}.")
-        properties['backend_domain'] = expected_backend_domain
-
-        if properties["ident"] != expected_ident:
-            raise UnexpectedDeviceProperty(
-                f"Got device with id: {properties['ident']} "
-                f"when expected id: {expected_ident}.")
-
-        if expected_devclass and properties['devclass'] != expected_devclass:
-            raise UnexpectedDeviceProperty(
-                f"Got {properties['devclass']} device "
-                f"when expected {expected_devclass}.")
-
-        properties['attach_automatically'] = qbool(
-            properties['attach_automatically'])
-        properties['required'] = qbool(properties['required'])
-
-        return cls(**properties)
+import qubesadmin.exc
+from qubesadmin.device_protocol import (Device, DeviceInfo, UnknownDevice,
+                                        DeviceAssignment)
 
 
 class DeviceCollection:
@@ -959,7 +110,7 @@ class DeviceCollection:
         if not assignment.devclass_is_set:
             assignment.devclass = self._class
         elif assignment.devclass != self._class:
-            raise qubesadmin.ext.QubesValueError(
+            raise qubesadmin.exc.QubesValueError(
                 f"Device class does not match to expected: "
                 f"{assignment.devclass=}!={self._class=}")
 
@@ -1094,7 +245,7 @@ class DeviceCollection:
                 return dev
         # if still nothing, return UnknownDevice instance for the reason
         # explained in docstring, but don't cache it
-        return UnknownDevice(self._vm, self._class, item)
+        return UnknownDevice(self._vm, item, devclass=self._class)
 
 
 class DeviceManager(dict):
